moved scroll button handler to helpers
